Fix dispatching enqueued events when react instance may get recreated
The main problem with handling react instance recreation was that EventEmitter which is a singleton was storing enabled flag that should be reset whenever react instance goes down.
With this change we no longer store global state in emitter except from the events queue. When the even comes in we verify with the UAReactModule instance that the events should be delivered (new module object is instantiated for each context so it is safe to store such a state there).

Lastly there was another global reference being stored in EventEmitter and it was for react instance manager. As it was only there to cache the reference instead of getting it via application context it was a simple change to remove it. Retaining the instance was also problematic as the context manager may get recreated in apps that use native navigation (e.g. react-native-navigation library).